// operations-timeline/src/generated/models/cr2b6_batchesesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface cr2b6_batcheses {
  cr2b6_activecompoundg?: number;
  cr2b6_activecompoundg_date?: Date;
  cr2b6_activecompoundg_state?: number;
  cr2b6_batchesid: string;
  cr2b6_batchnumber: string;
  cr2b6_completedon?: Date;
  cr2b6_homogenizerstarttime?: Date;
  cr2b6_homogenizerstarttime_date?: Date;
  cr2b6_homogenizerstarttime_state?: number;
  cr2b6_liquidvolumel?: number;
  cr2b6_liquidvolumel_date?: Date;
  cr2b6_liquidvolumel_state?: number;
  cr2b6_mfstarttime?: Date;
  cr2b6_mfstarttime_date?: Date;
  cr2b6_mfstarttime_state?: number;
  cr2b6_notes?: string;
  cr2b6_product?: string;
  cr2b6_products?: string;
  cr2b6_productsname?: string;
  cr2b6_purpose?: string;
  cr2b6_purposename?: string;
  cr2b6_recoverycentrifuge?: number;
  cr2b6_recoveryhomogenizer?: number;
  cr2b6_recoverymf?: number;
  cr2b6_recoverypredrying?: number;
  cr2b6_recoveryuf?: number;
  cr2b6_solidsmassg?: number;
  cr2b6_solidsmassg_date?: Date;
  cr2b6_solidsmassg_state?: number;
  cr2b6_trialplan?: string;
  cr2b6_trialplan_name?: string;
  cr2b6_ufstarttime?: Date;
  cr2b6_ufstarttime_date?: Date;
  cr2b6_ufstarttime_state?: number;
  createdby?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: Date;
  createdonbehalfby?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  importsequencenumber?: number;
  modifiedby?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: Date;
  modifiedonbehalfby?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  overriddencreatedon?: Date;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunit?: string;
  owningbusinessunitname: string;
  owningteam?: string;
  owninguser?: string;
  statecode: string;
  statecodename?: string;
  statuscode?: string;
  statuscodename?: string;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
  versionnumber?: number;
}